Try blocktiled sparse fwd

+++ src/sparse_fwd/1_naive.cu
@@ -1,6 +1,6 @@
-__global__ void sparse_fwd_naive_kernel(const size_t m, const size_t nnz, const float* A, const int32_t* x, float* y)
+__global__ void sparse_fwd_naive_kernel(const int32_t m, const int32_t nnz, const float* A, const int32_t* x, float* y)
 {
-    const int row = threadIdx.x + blockDim.x * blockIdx.x;
+    const int32_t row = threadIdx.x + blockDim.x * blockIdx.x;
     const int32_t* tx = x + nnz * blockIdx.y;
     float* ty = y + m * blockIdx.y;
